refactor(cli): consolidate api infra files (E2)
Intra-package moves, no symbol or behavior changes:
- errors.go absorbs sdkerr.go (HTTPError/apiErrorFor/deref/ptr/
  strEmptyToNil) plus reportCoded/asCoded from context_helpers.go.
- delete_cmd.go takes deleteSpec/newResourceDeleteCmd/article.
- common.go splits into session.go (agentSession/requireState/
  controlClient/contextSession + coded auth errors) and args.go
  (exactNamedArgs/rangeNamedArgs/mustMarkRequired).
- context_helpers.go and sdkerr.go removed once emptied.
No guarded call moved (E0). curated/migrate/legacy_store untouched.
